Move more hero images into the CMS
