+++ config/locales/en.yml
@@ -307,8 +307,8 @@ toplevel:
           - "Offering long-term, flexible funding and support to charities and community groups"
           - "Helping organisations to work together with others to improve communities and solve problems"
   ebulletin:
-    title: Find out about our funding
-    shortTitle: e-bulletin
+    title: e-bulletin
+    introHeading: Find out about our funding
     intro: Sign up for our latest news relating to your country
     location: Location
     locations:
@@ -401,7 +401,7 @@ toplevel:
           and research (PDF 1.8MB)
         href: https://www.biglotteryfund.org.uk/-/media/Files/Research%20Documents/er_res_answering_big_questions.pdf
   data:
-    title: 'Key data about the Big Lottery Fund: <br />financial year 2016/17'
+    title: "Key data about the Big Lottery Fund: financial year 2016/17"
     keyStats: Key statistics
     map:
       title: Our Year in Numbers
