+++ views/components/ebulletin.njk
@@ -14,7 +14,7 @@
 
 {% macro ebulletinForm() %}
 <form class="form-ebulletin" method="post" action="/about/ebulletin">
-    <h2 class="t1 t--underline accent--pink a--text">{{ __('toplevel.ebulletin.title') }}</h2>
+    <h2 class="t1 t--underline accent--pink a--text">{{ __('toplevel.ebulletin.introHeading') }}</h2>
 
     {% if status === formHelpers.FORM_STATES.SUBMISSION_SUCCESS %}
         {{ submissionMessage(
