+++ views/pages/funding/programmes.njk
@@ -4,11 +4,6 @@
 
 {% extends "layouts/main.njk" %}
 
-{% block title %}{{ title }} | {% endblock %}
-
-{% set heroImage = heroImages.youngFoundation %}
-{% set socialImage = heroImage %}
-
 {% block content %}
     {% set pageAccent = 'pink' %}
     <main role="main">
