+++ views/pages/toplevel/styleguide.njk
@@ -132,8 +132,8 @@
 
 {% set sgContent %}
     {{ hero(
-        titleText = 'Funding Programmes',
-        image = heroImages.youngFoundation,
+        titleText = 'Hero title text',
+        image = heroImages.rathlinIslandDevelopment,
         accent = 'green'
     ) }}
 {% endset %}
